requirements.txt
asgiref==3.2.10
astroid==2.4.2
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
defusedxml==0.7.0rc1
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2
django-cors-headers==3.4.0
django-extensions==3.0.9
django-filter==2.3.0
django-heroku==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
djoser==2.0.3
Fabric3==1.14.post1
geographiclib==1.50
geopy==2.0.0
gunicorn==20.0.4
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
paramiko==2.7.1
Pillow==7.2.0
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
PyNaCl==1.4.0
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
static3==0.7.0
toml==0.10.1
urllib3==1.25.10
whitenoise==5.2.0
wrapt==1.12.1
drf-extra-fields==3.0.3
django_rest_swagger==2.2.0
drf-yasg==1.17.1